added generate answer method for AI and refactored
