+++ story-ai/src/main/java/model/User.java
@@ -1,0 +1,4 @@
+package model;
+
+public class User {
+}
