+++ story-ai/src/main/java/model/story/Character.java
@@ -1,4 +1,4 @@
-package model;
+package model.story;
 
 public class Character {
