+++ story-ai/src/main/java/model/story/PromptBuilder.java
@@ -1,4 +1,4 @@
-package model;
+package model.story;
 
 import java.util.List;
